+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## 1.0.8 - 2021-09-04
+
+* New: Drag and drop taxonomy order
+* Changed: Restructure and cleanup code
+
 ## 1.0.7 - 2021-08-29
 
 * Fixed: Skip convert plain text to email address if content has input type email address